Begin tracking changes in entity descendents.
An entity must be considered "modified" if any of its descendents are
also modified.  In Git at least, the modification flag tells us
whether an entity can become a tag.  Unmodified branches are tags, and
modified ones are not.  However, a branch cannot be made of a tag (but
a tag can be renamed).  So if descendents of tags are modified, it
means that the tags themselves cannot be tags.