Pre-emptively remove files that have been git-rm'd but not committed.
Git-rm doesn't always remove files right away.  Snerp Vortex is
strict, however.  For sanity's sake, files are expected to be gone
right when they are deleted.  If a file is added or copied onto an
existing one, that's a fatal error.  So we'll just delete the files
after git-rm'ing them.  It seems harmless from git's POV.