Add a --prefix argument.
This is passed onwards to the SVN dump walker.  It will cause the
walker to skip any SVN::Dump records with a Node-path that doesn't
match this prefix.  Snerp Vortex will bail out with an error if files
are copied into the prefix from outside the prefix, since this would
cause a loss of svn history.