Reimplement on_node_replace in terms of delete & add.  Experimental.
At least one big Subversion dump I've seen contains a replace action
for a directory node.  Considering it a "change" does it no justice,
since all the subdirectories remain in the analysis.  It must be
analyzed as a delete/add pair, or we'll need a special "replace"
action.  This change implements the delete/add pair.  It's
experimental and untested until I can find or make a small dump file
that reproduces the offending directory replacement.