Calculate md5 summaries for new snubbed text.
svnadmin load checks the MD5 summaries for text and rejects the dump
if they don't match what's actually in the file.  Still to do,
however, are the SHA1 summaries (if present) and the summaries for
copy sources.  Some of them are compensated for, but there are still
edge cases.  Shouldn't be a problem for the snerp toolchain, but it's
an issue when loading into svn for comparison.